requirements.txt
# Core dependencies for LangChain and Ollama (Updated for Python 3.13)
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0

# Vector databases
chromadb>=1.0.0
faiss-cpu==1.8.0.post1

# Document processing
pypdf==4.3.1

# Text processing and embeddings
sentence-transformers==3.0.1
transformers==4.45.2

# Utilities and system
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
psutil==6.0.0

# Logging and debugging
colorlog==6.8.2

# Authentication and security
passlib==1.7.4
bcrypt==4.1.3  # Compatible con passlib 1.7.4 (bcrypt 5.x tiene problemas de compatibilidad)
python-multipart==0.0.9

# Database
mysqlclient==2.2.4

# Optional dependencies for better performance
torch==2.4.1
tokenizers==0.20.0

# System dependencies (Updated)
# sqlite3-binary==0.5.2  # Not needed for Python 3.13
# pysqlite3==0.5.2       # Not needed for Python 3.13

# Development and testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0

# === ASYNC MIGRATION DEPENDENCIES ===
# Phase 0: Async Infrastructure
redis>=4.5.2,<6.0.0
celery==5.4.0
celery[redis]==5.4.0

# Database connectors
pymysql==1.1.1
cryptography==42.0.8
sqlalchemy==2.0.32
alembic==1.13.2

# Phase 1: FastAPI Async Support  
uvicorn[standard]>=0.30.0
fastapi>=0.115.0
sse-starlette==2.1.3

# Phase 2: High-Performance Inference (vLLM)
# Note: vLLM will be installed separately as it requires specific CUDA setup
# vllm==0.6.2  # Uncomment after CUDA verification

# Phase 3: Model Optimization
# torch-audio==2.4.1  # Already have torch above
accelerate==0.34.2
bitsandbytes==0.44.1

# Phase 4: Streaming and Real-time Features
websockets==13.1
aioredis==2.0.1

# Monitoring and Performance
flower==2.0.1  # Celery monitoring
# redis-py-cluster is obsolete and incompatible with modern redis versions